+++ pdr_backend/util/contract.py
@@ -11,6 +11,10 @@ from enforce_typing import enforce_types
 
 @enforce_types
 def get_address(web3_pp, contract_name: str):
+    """
+    Returns address for this contract
+    in web3_pp.address_file, in web3_pp.network
+    """
     network = get_addresses(web3_pp)
     if not network:
         raise ValueError(f"Cannot figure out {contract_name} address")
@@ -19,7 +23,10 @@ def get_address(web3_pp, contract_name: str):
 
 
 @enforce_types
-def get_addresses(web3_pp):
+def get_addresses(web3_pp) -> Union[dict, None]:
+    """
+    Returns addresses in web3_pp.address_file, in web3_pp.network
+    """
     address_file = web3_pp.address_file
 
     path = None
@@ -35,25 +42,13 @@ def get_addresses(web3_pp):
     with open(path) as f:
         d = json.load(f)
 
-    d = _saphire_to_sapphire(d)
+    d = _condition_sapphire_keys(d)
 
     if "barge" in web3_pp.network:  # eg "barge-pytest"
         return d["development"]
     if web3_pp.network in d:  # eg "development", "oasis_sapphire"
         return d[web3_pp.network]
     return None
-
-
-@enforce_types
-def _saphire_to_sapphire(d: dict) -> dict:
-    """Fix spelling error coming from address.json"""
-    d2 = copy.deepcopy(d)
-    names = list(d.keys())  # eg ["mumbai", "oasis_saphire"]
-    for name in names:
-        if "saphire" in name:
-            fixed_name = name.replace("saphire", "sapphire")
-            d2[fixed_name] = d[name]
-    return d2
 
 
 @enforce_types
@@ -96,3 +91,19 @@ def get_contract_filename(contract_name: str, address_file: Union[str, None]):
     if not path.exists():
         raise TypeError(f"Contract '{contract_name}' not found in artifacts.")
     return path
+
+
+@enforce_types
+def _condition_sapphire_keys(d: dict) -> dict:
+    """
+    For each sapphire-related key seen from address.json,
+    transform it to something friendly to pdr-backend (and named better)
+    """
+    d2 = copy.deepcopy(d)
+    names = list(d.keys())  # eg ["mumbai", "oasis_saphire"]
+    for name in names:
+        if name == "oasis_saphire_testnet":
+            d2["sapphire-testnet"] = d[name]
+        elif name == "oasis_saphire":
+            d2["sapphire-mainnet"] = d[name]
+    return d2
